backend/pyproject.toml
[project]
name = "mianshibao-backend"
version = "0.1.0"
description = "Mianshibao backend service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.36",
    "psycopg[binary]>=3.2.3",
    "asyncpg>=0.29.0",
    "alembic>=1.13.3",
    "pgvector>=0.3.6",
    "passlib[bcrypt]>=1.7.4",
    "dashscope>=1.20.14",
    "tenacity>=9.0.0",
    "circuitbreaker>=2.0.0",
    "langchain-core>=0.3.17",
    "pdfplumber>=0.11.4",
    "python-docx>=1.1.2",
    "paddlepaddle>=3.0.0",
    "paddleocr>=2.8.1",
    "python-multipart>=0.0.20",
    "redis>=5.2.0",
    "minio>=7.2.10",
    "PyJWT>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pre-commit>=4.0.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
